requirements.txt
bert-embedding==1.0.1
scikit-learn==0.23.1
spacy==2.3.2
Keras==2.8.0
tensorflow==2.2.0
pandas==1.3.5
numpy==1.18.5
google-trans-new==1.1.9
langdetect==1.0.9
PyYAML==6.0
regex==2022.4.24
